Refactors dialogs for improved reusability
Generalizes the dialog base component to accept custom cancel and confirm texts via props, allowing for more flexible and reusable dialogs.

This change also updates the delete stock item dialog to utilize the common cancel and delete translations and the new dialog base props, resulting in a cleaner and more maintainable codebase. Furthermore, it adds a "create" translation to the common translations file.